### 2019-05-12 #1444 @whedon
[REVIEW]: modelDown: automated website generator with interpretable documentation for predictive machine learning models
**Submitting author:** @matiszak (<a href="http://orcid.org/0000-0001-6335-3434">Mateusz Urbański</a>)
**Repository:** <a href="https://github.com/MI2DataLab/modelDown" target ="_blank">https://github.com/MI2DataLab/modelDown</a>
**Version:** v1.0.1
