### 2019-05-13 03ff0df
Add OSX to travis check
