+++ DESCRIPTION
@@ -32,7 +32,6 @@ Suggests: ranger,
     testthat,
     drifter,
     useful
-Additional_repositories: https://github.com/ModelOriented/drifter
 RoxygenNote: 6.1.1
 URL: https://github.com/MI2DataLab/modelDown
 BugReports: https://github.com/MI2DataLab/modelDown/issues
