### 2019-05-13 2079343
Update authors section
