+++ DESCRIPTION
@@ -1,7 +1,8 @@
 Package: modelDown
 Title: Make Static HTML Website for Predictive Models
 Version: 1.0.0
-Authors@R: c(person("Magda", "Tatarynowicz", email = "magda.tatarynowicz@gmail.com", role = c("aut")), 
+Authors@R: c(person("Przemysław", "Biecek", email = "przemyslaw.biecek@gmail.com", role = c("aut")), 
+    person("Magda", "Tatarynowicz", email = "magda.tatarynowicz@gmail.com", role = c("aut")), 
     person("Kamil", "Romaszko", email = "kamil.romaszko@gmail.com", role = c("aut", "cre")), 
     person("Mateusz", "Urbański", email = "matiszak@gmail.com", role = c("aut")))
 Description: The package 'modelDown' generates a website with HTML summaries for predictive models.
@@ -32,7 +33,6 @@ Suggests: ranger,
     testthat,
     drifter,
     useful
-Additional_repositories: https://github.com/ModelOriented
 RoxygenNote: 6.1.1
 URL: https://github.com/MI2DataLab/modelDown
 BugReports: https://github.com/MI2DataLab/modelDown/issues
