### 2019-05-15 0639ac0
Update DESCRIPTION and add cran-comments.md
