### 2019-05-15 bf30d46
Update cran-comments.md
