### 2019-05-15 6963a41
Update comments to NOTEs in cran-comments.md
