### 2019-05-27 #1444 @pdwaggoner
Hi - Sorry for the delay. Overall, this looks really good! Here are a few comments for things I think should be addressed/fixed:

1. The license on the main repo (Apache 2.0) doesn't match the license at the pkgdown "documentation" site (GPL-3 -- https://mi2datalab.github.io/modelDown/). Both are acceptable by OSI, but I would clarify and fix.

2. Installation instructions are included in the vignette, but I would like to see something a little more clear and up front on the repo README landing (e.g., `Installation for Dev Version: xxx // Installation for Stable Version: xxx`). Something simple and straightforward for the user would be a useful place to start interaction with the package.

3. Typo on page 1 in the paragraph starting with "The modelDown package is designed..." in the sentence, "In THE case of pkgdown... while in THE case of modelDown..."

4. Typo on page 2, in the paragraph starting with "The 'Model performance'..." -- the word "lab" should be "tab". There are many other small grammatical errors/typos throughout. I would recommend a closer second (or third) read-through of the paper to root these out and correct them. Should be a simple fix.

5. I don't think the first sentence of the Summary section (which is slightly grammatically incorrect) is too useful.

6. I am not sure if the authors were intending to include a sample screenshot or figure in the two places with hyphens (e.g., "-Example model performance tab."), but if not, I would highly recommend including a screenshot to give the user/reader an idea of what the output looks like. If so, then perhaps the authors forgot to include this in the manuscript. I would include some demo figures either way.

7. An example/demo of the code would be a helpful addition to the paper. Perhaps just before the explanation of all of the tabs produced in the output. It would be nice just to see the code/package in action within the JOSS paper, even though examples are included in the package vignette.

8. I would also include a note on how people can contribute (if you want them to), via issue tickets, PRs, or whatever this might look like.

Other than these things, the paper and software all looks great! The authors did a good job here and I think the package will be widely accepted and used. Great work!
