### 2019-06-01 48ff728
Update CRAN comments
