+++ cran-comments.md
@@ -11,11 +11,9 @@ My first cran submission. Here are comments to check NOTEs:
 
 There were no ERRORs or WARNINGs
 
-There were 2 NOTEs:
+There were 1 NOTEs:
 
 * Possibly mis-spelled words in DESCRIPTION:
     * NCN (14:59) - It's polish abbreviation for "Narodowe Centrum Nauki"
-    * explainers (8:31) - It's proper name introduced in DALEX package. It describes object that contains ML model and additional information for it
+    * explainers (8:31) - It's proper name introduced in DALEX package. It describes object that contains ML model with addiitional meta informations
 
-* Suggests or Enhances not in mainstream repositories (also suggested but not available for checking):
-    * drifter - package is part of our organisation and is not published to cran yet. Repository could be found here: https://github.com/ModelOriented/drifter. It's not required for main functionality of modelDown. 
