### 2019-06-02 fa89d94
Add package (drifter) to DESCRIPTION to avoid errors/warnings
