### 2019-06-02 1642590
Add example to paper.md
