### 2019-06-02 8a959e0
Bump version to 1.0.1
