### 2019-06-02 32208bb
README: Add Installation & Contibuting info
