### 2019-06-02 e8e7558
README: Add contributing guide (exact steps)
