### 2019-06-02 f655aa6
paper.md: grammatical corrections
