+++ paper.md
@@ -28,17 +28,17 @@ bibliography: paper.bib
 
 We live in times where machine learning is used almost everywhere. With the amount of accessible tools available, generation of predictive model for your own benefit is fairly straightforward. 
 
-The models may be simple to create and use, but most of the time they are not simple to understand. Large models may contain thousands or millions of parameters. It can be challenging to see how the model works and how the inputs impacts obtained results. Numerous tools for eXplainable Artificial Intelligence can help with that, but it may be cumbersome and repetitive to use different tools every time you test the model. 
+The models may be simple to create and use, but most of the time they are not simple to understand. Larger models may contain thousands or millions of parameters. It can be challenging to see how the model works and how the inputs impacts obtained results. Numerous tools for explainable Artificial Intelligence can help with that, but it may be cumbersome and repetitive to use different tools every time you test the model. 
 
-In this paper we introduce a package modelDown for the R language [@R] that can help to eliminate these issues. It automatically creates an easy to understand HTML website with model interpretations. You can automatically analyze, explore and debug your predictive models in many ways with just a single command.
+In this paper, we introduce a package modelDown for the R language [@R] that can help to eliminate these issues. It automatically creates an easy to understand HTML website with model interpretations. You can automatically analyze, explore and debug your predictive models in many ways with just a single command.
 
-The modelDown package is designed in a similar way to the pkgdown [@pkgdown] library. Both packages use templates to generate a static HTML website with detailed information. In case of pkgdown it is an information about a selected package, while in case of modelDown it is a website with information about predictive models.
+The modelDown package is designed in a similar way to the pkgdown [@pkgdown] library. Both packages use templates to generate a static HTML website with detailed information. In the case of pkgdown it is an information about a selected package, while in the case of modelDown it is a website with information about predictive models.
 
 The modelDown package builds on the set of tools for model exploration implemented in the DALEX [@DALEX] package. By creating explainer objects, DALEX provides uniform interface for model agnostic exploration of models. Other R packages compatible with DALEX are also used to provide data, tables and charts available on the website.
 
 # Example code
 
-Basic example of using the package is as straightforward as below.
+A basic example of using the package is as straightforward as below.
 
 ```
 require("ranger")
@@ -72,36 +72,35 @@ modelDown::modelDown(explainer_ranger, list(explainer_glm1, explainer_glm2))
 
 # Available features
 
-The modelDown package is highly modular. The primary outcome is a website with set of tabs. Each tab explains a particular aspect of the model. Features are available across tabs. By default, all tabs are generated, and each of them can be parameterized. 
+The modelDown package is highly modular. The primary outcome is a website with a set of tabs. Each tab explains a particular aspect of the model. Features are available across the tabs. By default, all the tabs are generated, and each of them can be parameterized. 
 
-The 'Auditor' tab presents different ways of validating predictive models by analysis of residuals. All plots are generated with the auditor package [@Auditor]. The plots may seem a bit complicated at first sight, so each plot has links with helpful explanations. 
+The 'Auditor' tab presents different ways of validating predictive models by analysis of residuals. All plots are generated by the auditor package [@Auditor]. The plots may seem a bit complicated at first sight, so each plot has links with helpful explanations. 
 
 The 'Model performance' tab uses functions from DALEX package [@DALEX]. It shows contrastive comparisons of model predictive power by analyzing residuals. Read more about this type of model diagnostic in [@PM_VEE].
 
-The 'Model performance' tab lets you compare different models. The same applies for all tabs created with modelDown. This allows user not only to validate one model, but to pick the best model among many. Below we can see that the best ROC curve is obtained for Random Forest model.
[…]
