### 2019-06-02 15b5f22
Update JOSS paper_images
