### 2019-06-02 #1444 @Matiszak
Thank you very much for your review and sorry for the delay. I hope I took care of all the points, your input was very helpful. Could you see if there is anything else worth changing ?
