### 2019-06-03 #1444 @terrytangyuan
Just went over the paper and the checklist. The paper looks great to me. Great job @Matiszak!
