### 2019-06-12 47572f7
Make example simple and fast
