### 2019-06-13 #1444 @arfon
> @Matiszak do you have an archive DOI for the repository? We need that for acceptance.

Also, @Matiszak - please make sure that you make new release of this software that includes the changes that have resulted from this review. Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive. For the Zenodo/figshare archive, please make sure that:
