### 2019-06-14 d7ad6d9
Add more complex example and update man doc
