### 2019-06-14 e56635b
Fix whitespaces in doc
