### 2019-06-17 3f91b73
Fix typo in an author name
