### 2019-06-17 #1444 @Matiszak
The 1.0.1 version including the changes that have resulted from this review has been archived at 10.5281/zenodo.3247303
Also it set as official release on https://github.com/MI2DataLab/modelDown and also published to CRAN: https://cran.r-project.org/web/packages/modelDown/index.html

Thank you very much for your work, please let me know if there is anything else to do.
