### 2019-06-18 #1444 @alexhanna
@openjournals/joss-eics this looks ready for acceptance.
