### 2019-06-18 ee6fe43
a bunch of small changes in the paper text
