### 2019-06-18 #1444 @danielskatz
I've suggested a bunch of text changes for the paper in https://github.com/MI2DataLab/modelDown/pull/72
