### 2019-06-25 #1444 @danielskatz
👋 @Matiszak - please also merge https://github.com/MI2DataLab/modelDown/pull/74, which has fixes for the cases of words in the bibtex. Then ping @openjournals/joss-eics for the final acceptance
