### 2019-06-27 #1444 @Matiszak
@openjournal/joss-eics Merged :+1:
