### 2019-06-27 #1444 @Matiszak
@openjournals/joss-eics Merged :+1:
